+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- When copying a file from an existing GCS object, the content type was not properly copied. This is a bug in the `google-cloud-storage` gem, a workaround has been added so it now works currently (issue #36)
+
 ## 2.0.0 - 2019-02-12
 
 ### Added